Changed to foreground service to improve reliability

+++ app/src/main/java/com/anthonyla/paperize/feature/wallpaper/presentation/home_screen/HomeScreen.kt
@@ -38,7 +38,7 @@ fun HomeScreen(
     onContactClick: () -> Unit,
     navigateToAddWallpaperScreen: (String) -> Unit,
     onAlbumViewClick: (String) -> Unit,
-    onScheduleWallpaperChanger: (Long) -> Unit
+    onScheduleWallpaperChanger: (Int) -> Unit
 ) {
     var tabIndex by rememberSaveable { mutableIntStateOf(0) }
     val pagerState = rememberPagerState { tabItems.size }
